add background image in ChallengeCard
