+++ components/homepage/css/ChallengeCard.module.css
@@ -1,10 +1,4 @@
 .ChallengeCard {
-
-    background: var(--TextDimmedDark);
-    display: flex;
-    flex-direction: row;
-    justify-content: space-between;
-    align-items: flex-end;
     padding: 10px 17px;
     gap: 124px;
 
@@ -13,17 +7,20 @@
 
     border-radius: 8px;
 
-    flex: none;
-    order: 0;
-    flex-grow: 0;
+    transition: 0.2s;
+    transition-timing-function:ease-out;
+}
+
+.ChallengeCard:hover {
+    transform: scaleX(1.01) scaleY(1.02);
 }
 
 .BottomHalf {
-    display: flex;
+    /* display: flex;
     flex-direction: column;
     align-items: flex-start;
     padding: 0px;
-    gap: 9px;
+    gap: 9px; */
 
     width: 100%;
     height: 95px;
@@ -134,4 +131,11 @@
 .Icon {
     width: 20px;
     height: 20px
+}
+
+.Img {
+    z-index: -1;
+    opacity: 0.8;
+    transition: 0.2s ease-in;
+    filter: brightness(47.5%)
 }
