+++ components/homepage/css/ChallengeDashboard.module.css
@@ -1,10 +1,6 @@
 .Challenges {
     padding: 30px 15px 10px;
     gap: 8px;
-
-
     width: 1000px;
-
-    box-shadow: 0px 2px 8px rgba(0, 0, 0, 0.25);
     border-radius: 10px;
 }
