+++ components/homepage/css/TaskDashboard.module.css
@@ -1,12 +1,7 @@
 .TaskDashboard {
     padding: 30px 15px 10px;
-
-    width: 250px;
+    width: 275px;
     height: 100%;
-
-    background: var(--White);
-
-    box-shadow: 0px 2px 8px rgba(0, 0, 0, 0.25);
     border-radius: 10px;
 }
