+++ pages/home/index.tsx
@@ -1,3 +1,19 @@
-import HomePage from '../index'
+import Head from 'next/head'
+import ChallengeDashboard from '../../components/homepage/ChallengeDashboard';
+import TaskDashboard from '../../components/homepage/TaskDashboard';
 
-export default HomePage
+
+export default function Home() {
+
+  return (
+    <>
+      <Head>
+        <title>Home</title>
+      </Head>
+      <div className='flex justify-center items-start space-x-4 mt-6'>
+        <ChallengeDashboard></ChallengeDashboard>
+        <TaskDashboard></TaskDashboard>
+      </div>
+    </>
+  );
+}
