+++ styles/globals.css
@@ -111,3 +111,10 @@
     font-size: 12px;
 }
 
+.ShadowContainer {
+    box-shadow: 0px 2px 8px rgba(0, 0, 0, 0.25);
+}
+
+.ShadowButton {
+    box-shadow: 0px 4px 4px rgba(0, 0, 0, 0.25);
+}
